Revision 2 for PR #11
Updated filenames to differentiate web from native
