+++ native/index.js
@@ -1,3 +1,3 @@
-const toHaveStyleRule = require('../src/matchers/toHaveStyleRule')
+const toHaveStyleRule = require('../src/matchers/toHaveStyleRule.native')
 
 expect.extend({ toHaveStyleRule })
