+++ src/index.js
@@ -1,5 +1,5 @@
 const toMatchStyledComponentsSnapshot = require('./matchers/toMatchStyledComponentsSnapshot')
-const toHaveStyleRule = require('./matchers/toHaveStyle')
+const toHaveStyleRule = require('./matchers/toHaveStyleRule')
 const styleSheetSerializer = require('./serializers/styleSheetSerializer')
 
 expect.addSnapshotSerializer(styleSheetSerializer)
